+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 framer motion bottom sheet
 
+## [0.1.3] 2023-07-03
+
+### Feature
+
+- add Props FooterElement
+- rename Sheet.tsx -> FramerBottomSheet.tsx
+- add example TestPage
+
 ## [0.1.2] 2023-07-03
 
 ### Fixed